--- src/components/organisms/Post/TagList/TagList.type.ts
@@ -1,6 +1,0 @@
-import { CSSProperties } from 'react';
-
-export type TagListProps = {
-  tags: string[];
-  style?: CSSProperties;
-};
