--- src/components/organisms/Post/TagList/useTagList.ts
@@ -1,9 +1,0 @@
-import { useTheme } from '@emotion/react';
-
-export const useTagList = () => {
-  const theme = useTheme();
-
-  return {
-    theme,
-  };
-};
